renaming shortify package to app
