+++ app/database_test.go
@@ -1,4 +1,4 @@
-package shortify
+package app
 
 import (
 	"github.com/stretchr/testify/assert"
