+++ app/handlers_test.go
@@ -1,4 +1,4 @@
-package shortify
+package app
 
 import (
 	"bytes"
