+++ app/user.go
@@ -1,4 +1,4 @@
-package shortify
+package app
 
 import (
 	"code.google.com/p/go-uuid/uuid"
